Разбиваем на модули js
